Study Note for Feature Engineering
There was ambiguous points in the article I studied. First of all, it did not explain why apparent temperature gives better information than individual variables like temperature. Second, It did not confirm that the error was reduced simply by increasing the discrimination. Since the principle of decision tree is repeatedly finding the discrimination, the error usually decreases if new feature is added, whether it is meaningful or not. I filled these two points in this file.